Made skills logos horizontal on smaller screens; fixed big-pictures on smaller screens
On smaller screens, the logos in the Technical Skills sections would
just be stacked, 1 per row. It worked, but it took a while to scroll
through. Now their sizes are just reduced and they are stacked
horizontally, so they take up 1/3 of the space.

Also, big-pictures were sized too small on smaller screens. Their
border-radii were not adjusted properly in accordance with their
resizing.